+++ packages/pc/src/graph/controller/layout.ts
@@ -490,7 +490,7 @@ export default class LayoutController extends AbstractLayout {
 
     let start = Promise.resolve();
     let hasLayout = false;
-    if (layoutMethods.length === 1) {
+    if (layoutMethods?.length === 1) {
       hasLayout = true;
       start = start.then(async () => await this.updateLayoutMethod(layoutMethods[0], layoutCfg));
     } else {